refactor(cli): split cli.py into a package by concern
cli.py had grown to 1,902 lines and was the most-churned source file in the
repo — +21% across the last four releases, which is what makes it worth
splitting ahead of larger but stable files. It held eight Typer commands plus
report rendering, SARIF/HTML/badge generation, language-server provisioning,
trust filters and onboarding.

An AST pass over the candidate groups showed they were already nearly
independent: apart from `console`, `err_console` and `app`, each needed almost
nothing from the rest of the file. So the seams were there; only the shared
Rich consoles and the Typer app held them together.

  _app.py         the Typer app, in its own module so command modules can
                  register on it without importing the package __init__
  _io.py          the two shared consoles
  config.py       config dir + API key access
  environment.py  detection only — what is installed on this machine
  setup_lsp.py    the `setup` command, which acts on what environment finds
  main.py         the remaining seven commands (to be split further)

Detection was separated from action deliberately: the scan pre-flight and the
`setup` command both need _first_which/_chromium_installed/_LSP_SPECS, and a
scan path importing from a setup module would have been the wrong dependency.

__init__ is a façade exposing only `app` and the consoles. It does not
re-export the private helpers, because a name re-exported there could be
patched without affecting the module that actually uses it — a seam that
looks like it works and doesn't. Tests now patch where each name is used.

Two things this turned up that were worth fixing rather than shipping:
importing setup_lsp at main's module scope registered `setup` first and
silently reordered `--help`, so `launch` imports _lsp_offer lazily; and the
first cut left _ensure_config_dir/_load_api_key defined in two modules, now
shared via config.py.

32 of the 34 top-level functions moved byte-identical, none lost, none
duplicated. The other two are the lazy import in `launch` and the deduplicated
_load_api_key.

Verified beyond unit tests, since the entry point moved: benchmarks 46/46
taint recall with 0 FP and 3/3 on VAmPI; Juice Shop 198 findings in repo mode
and 26 live DAST findings with an unchanged category split; test-app 114; the
#147 bad-branch exit code still 1. A wheel built and installed clean into a
fresh venv resolves isitsecure.cli:app and lists all eight commands.

Release-As: 0.22.3
